UsuarioDTO + ID systax change

+++ BackEnd/repositories/AlergenicoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AlergenicoRepository extends JpaRepository<Alergenico, Integer> {
+public interface AlergenicoRepository extends JpaRepository<Alergenico, Long> {
 }
