+++ BackEnd/repositories/TabelaNutricionalRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TabelaNutricionalRepository extends JpaRepository<TabelaNutricional, Integer> {
+public interface TabelaNutricionalRepository extends JpaRepository<TabelaNutricional, Long> {
 }
